WIP #1 Created ClassGenerator as base to Model and Pipeline Generators; reorganized file structure
