--- generators/preprocessor/variable_map.json
@@ -1,6 +1,0 @@
-{
-    "variable_map" : {
-        "train_size" : "self.ds_info.splits['train'].num_examples",
-        "autotune" : "tf.data.experimental.AUTOTUNE"
-    }
-}
